+++ api/odl.py
@@ -1,82 +1,51 @@
 import datetime
-import dateutil
 import json
+import logging
 import uuid
-from flask_babel import lazy_gettext as _
-import urllib.parse
-from collections import defaultdict
-import flask
-from flask import Response
-import feedparser
-from lxml import etree
-from .problem_details import NO_LICENSES
 from io import StringIO
-import re
+from typing import Callable, Optional
+
+import dateutil
+import feedparser
+import flask
+from flask import url_for
+from flask_babel import lazy_gettext as _
+from lxml import etree
+from sqlalchemy.sql.expression import or_
 from uritemplate import URITemplate
 
-from sqlalchemy.sql.expression import or_
-
-from core.opds_import import (
-    OPDSXMLParser,
-    OPDSImporter,
-    OPDSImportMonitor,
-)
-from core.monitor import (
-    CollectionMonitor,
-    TimelineMonitor,
-)
+from core import util
+from core.analytics import Analytics
+from core.metadata_layer import CirculationData, FormatData, LicenseData, TimestampData
 from core.model import (
     Collection,
     ConfigurationSetting,
-    Credential,
     DataSource,
     DeliveryMechanism,
     Edition,
     ExternalIntegration,
     Hold,
     Hyperlink,
-    Identifier,
-    IntegrationClient,
     LicensePool,
     Loan,
     MediaTypes,
+    Representation,
     RightsStatus,
     Session,
-    create,
     get_one,
     get_one_or_create,
 )
-from core.metadata_layer import (
-    CirculationData,
-    FormatData,
-    IdentifierData,
-    LicenseData,
-    TimestampData,
-)
-from .circulation import (
-    BaseCirculationAPI,
-    LoanInfo,
-    FulfillmentInfo,
-    HoldInfo,
-)
-from core.analytics import Analytics
-from core.util.datetime_helpers import (
-    utc_now,
-    strptime_utc,
-)
-from core.util.http import (
-    HTTP,
-    BadResponseException,
-    RemoteIntegrationException,
-)
+from core.monitor import CollectionMonitor, IdentifierSweepMonitor
+from core.opds_import import OPDSImporter, OPDSImportMonitor, OPDSXMLParser
+from core.testing import DatabaseTest, MockRequestsResponse
+from core.util.datetime_helpers import utc_now
+from core.util.http import HTTP, BadResponseException, RemoteIntegrationException
 from core.util.string_helpers import base64
-from flask import url_for
-from core.testing import (
-    DatabaseTest,
-    MockRequestsResponse,
-)
+
+from .circulation import BaseCirculationAPI, FulfillmentInfo, HoldInfo, LoanInfo
 from .circulation_exceptions import *
 from .shared_collection import BaseSharedCollectionAPI
+
 
 class ODLAPI(BaseCirculationAPI, BaseSharedCollectionAPI):
     """ODL (Open Distribution to Libraries) is a specification that allows
@@ -596,7 +565,7 @@ class ODLAPI(BaseCirculationAPI, BaseSharedCollectionAPI):
                 Loan.end>utc_now()
             )
         ).count()
-        remaining_licenses = licensepool.licenses_owned - loans_count
+        remaining_licenses = max(licensepool.licenses_owned - loans_count, 0)
 
         holds = _db.query(Hold).filter(
             Hold.license_pool_id==licensepool.id
@@ -782,6 +751,7 @@ class ODLXMLParser(OPDSXMLParser):
     NAMESPACES = dict(OPDSXMLParser.NAMESPACES,
                       odl="http://opds-spec.org/odl")
 
+
 class ODLImporter(OPDSImporter):
     """Import information and formats from an ODL feed.
 
@@ -794,6 +764,105 @@ class ODLImporter(OPDSImporter):
     # The media type for a License Info Docuemnt, used to get information
     # about the license.
     LICENSE_INFO_DOCUMENT_MEDIA_TYPE = 'application/vnd.odl.info+json'
+
+    @classmethod
+    def parse_license(
+            cls,
+            identifier: str,
+            total_checkouts: Optional[int],
+            concurrent_checkouts: Optional[int],
+            expires: Optional[datetime.datetime],
+            checkout_link: Optional[str],
+            odl_status_link: Optional[str],
+            do_get: Callable
+    ) -> Optional[LicenseData]:
+        """Check the license's attributes passed as parameters:
+        - if they're correct, turn them into a LicenseData object
+        - otherwise, return a None
+
+        :param identifier: License's identifier
+        :param total_checkouts: Total number of checkouts before the license expires
+        :param concurrent_checkouts: Number of concurrent checkouts allowed for this license
+        :param expires: Date & time until the license is valid
+        :param checkout_link: License's checkout link
+        :param odl_status_link: License Info Document's link
+        :param do_get: Callback performing HTTP GET method
+
+        :return: LicenseData if all the license's attributes are correct, None, otherwise
+        """
+        remaining_checkouts = None
+        available_concurrent_checkouts = None
+
+        # This cycle ends in two different cases:
+        # - when at least one of the parameters is invalid; in this case, the method returns None.
+        # - when all the parameters are valid; in this case, the method returns a LicenseData object.
+        while True:
+            if total_checkouts is not None:
+                total_checkouts = int(total_checkouts)
+
+                if total_checkouts <= 0:
+                    logging.info(
+                        f"License # {identifier} expired since "
+                        f"the total number of checkouts is {total_checkouts}"
+                    )
+                    break
+
+            if expires:
+                if not isinstance(expires, datetime.datetime):
+                    expires = dateutil.parser.parse(expires)
+
+                expires = util.datetime_helpers.to_utc(expires)
+                now = util.datetime_helpers.utc_now()
+
+                if expires <= now:
+                    logging.info(
+                        f"License # {identifier} expired at {expires} (now is {now})"
+                    )
+                    break
+
+            if odl_status_link:
+                status_code, _, response = do_get(
+                    odl_status_link, headers={}
+                )
+
+                if status_code in (200, 201):
+                    status = json.loads(response)
+                    checkouts = status.get("checkouts", {})
+                    remaining_checkouts = checkouts.get("left")
+                    available_concurrent_checkouts = checkouts.get("available")
+                else:
+                    logging.warning(
+                        f"License # {identifier}'s Info Document is not available. "
+                        f"Status link failed with {status_code} code"
+                    )
+                    break
+
+            if remaining_checkouts is None:
+                remaining_checkouts = total_checkouts
+
+            if remaining_checkouts is not None:
+                remaining_checkouts = int(remaining_checkouts)
+
+                if remaining_checkouts <= 0:
+                    logging.info(
+                        f"License # {identifier} expired since "
+                        f"the remaining number of checkouts is {remaining_checkouts}"
+                    )
+                    break
+
+            if available_concurrent_checkouts is None:
+                available_concurrent_checkouts = concurrent_checkouts
+
+            return LicenseData(
+                identifier=identifier,
+                checkout_url=checkout_link,
+                status_url=odl_status_link,
+                expires=expires,
+                remaining_checkouts=remaining_checkouts,
+                concurrent_checkouts=available_concurrent_checkouts,
+            )
+
+        return None
 
     @classmethod
     def _detail_for_elementtree_entry(cls, parser, entry_tag, feed_url=None, do_get=None):
@@ -852,11 +921,6 @@ class ODLImporter(OPDSImporter):
 
             data['medium'] = medium
 
-            expires = None
-            remaining_checkouts = None
-            available_checkouts = None
-            concurrent_checkouts = None
-
             checkout_link = None
             for link_tag in parser._xpath(odl_license_tag, 'odl:tlink') or []:
                 rel = link_tag.attrib.get("rel")
@@ -875,32 +939,33 @@ class ODLImporter(OPDSImporter):
                     odl_status_link = attrib.get("href")
                     break
 
-            # If we found one, retrieve it and get licensing information about this book.
-            if odl_status_link:
-                ignore, ignore, response = do_get(odl_status_link, headers={})
-                status = json.loads(response)
-                checkouts = status.get("checkouts", {})
-                remaining_checkouts = checkouts.get("left")
-                available_checkouts = checkouts.get("available")
+            expires = None
+            total_checkouts = None
+            concurrent_checkouts = None
 
             terms = parser._xpath(odl_license_tag, "odl:terms")
             if terms:
+                total_checkouts = subtag(terms[0], "odl:total_checkouts")
                 concurrent_checkouts = subtag(terms[0], "odl:concurrent_checkouts")
                 expires = subtag(terms[0], "odl:expires")
-                if expires:
-                    expires = dateutil.parser.parse(expires)
 
-            licenses_owned += int(concurrent_checkouts or 0)
-            licenses_available += int(available_checkouts or 0)
+            license = cls.parse_license(
+                identifier,
+                total_checkouts,
+                concurrent_checkouts,
+                expires,
+                checkout_link,
+                odl_status_link,
+                do_get
+            )
 
-            licenses.append(LicenseData(
-                identifier=identifier,
-                checkout_url=checkout_link,
-                status_url=odl_status_link,
-                expires=expires,
-                remaining_checkouts=remaining_checkouts,
-                concurrent_checkouts=concurrent_checkouts,
-            ))
+            if not license:
+                continue
+
+            licenses_owned += int(license.remaining_checkouts or 0)
+            licenses_available += int(license.concurrent_checkouts or 0)
+
+            licenses.append(license)
 
         if not data.get('circulation'):
             data['circulation'] = dict()
@@ -913,6 +978,7 @@ class ODLImporter(OPDSImporter):
         data['circulation']['licenses_owned'] = licenses_owned
         data['circulation']['licenses_available'] = licenses_available
         return data
+
 
 class ODLImportMonitor(OPDSImportMonitor):
     """Import information from an ODL feed."""
@@ -959,11 +1025,12 @@ class ODLHoldReaper(CollectionMonitor):
         progress = TimestampData(achievements=message)
         return progress
 
+
 class MockODLAPI(ODLAPI):
     """Mock API for tests that overrides _get and _url_for and tracks requests."""
 
     @classmethod
-    def mock_collection(self, _db):
+    def mock_collection(cls, _db, protocol=ODLAPI.NAME):
         """Create a mock ODL collection to use in tests."""
         library = DatabaseTest.make_default_library(_db)
         collection, ignore = get_one_or_create(
@@ -973,7 +1040,7 @@ class MockODLAPI(ODLAPI):
             )
         )
         integration = collection.create_external_integration(
-            protocol=ODLAPI.NAME
+            protocol=protocol
         )
         integration.username = 'a'
         integration.password = 'b'
@@ -1040,6 +1107,25 @@ class SharedODLAPI(BaseCirculationAPI):
 
         self.base_url = collection.external_account_id
 
+    @staticmethod
+    def _parse_feed_from_response(response):
+        """Parse ODL (Atom) feed from the HTTP response.
+
+        :param response: HTTP response
+        :type response: requests.Response
+
+        :return: Parsed ODL (Atom) feed
+        :rtype: dict
+        """
+        response_content = response.content
+
+        if not isinstance(response_content, (str, bytes)):
+            raise ValueError("Response content must be a string or byte-encoded value")
+
+        feed = feedparser.parse(response_content)
+
+        return feed
+
     def internal_format(self, delivery_mechanism):
         """Each consolidated copy is only available in one format, so we don't need
         a mapping to internal formats.
@@ -1091,7 +1177,8 @@ class SharedODLAPI(BaseCirculationAPI):
                 hold_info_response = self._get(hold.external_identifier)
             except RemoteIntegrationException as e:
                 raise CannotLoan()
-            feed = feedparser.parse(str(hold_info_response.content))
+
+            feed = self._parse_feed_from_response(hold_info_response)
             entries = feed.get("entries")
             if len(entries) < 1:
                 raise CannotLoan()
@@ -1117,7 +1204,8 @@ class SharedODLAPI(BaseCirculationAPI):
         elif response.status_code == 404:
             if hasattr(response, 'json') and response.json().get('type', '') == NO_LICENSES.uri:
                 raise NoLicenses()
-        feed = feedparser.parse(str(response.content))
+
+        feed = self._parse_feed_from_response(response)
         entries = feed.get("entries")
         if len(entries) < 1:
             raise CannotLoan()
@@ -1181,7 +1269,8 @@ class SharedODLAPI(BaseCirculationAPI):
             raise CannotReturn()
         if response.status_code == 404:
             raise NotCheckedOut()
-        feed = feedparser.parse(str(response.content))
+
+        feed = self._parse_feed_from_response(response)
         entries = feed.get("entries")
         if len(entries) < 1:
             raise CannotReturn()
@@ -1286,7 +1375,8 @@ class SharedODLAPI(BaseCirculationAPI):
             raise CannotReleaseHold()
         if response.status_code == 404:
             raise NotOnHold()
-        feed = feedparser.parse(str(response.content))
+
+        feed = self._parse_feed_from_response(response)
         entries = feed.get("entries")
         if len(entries) < 1:
             raise CannotReleaseHold()
@@ -1325,7 +1415,7 @@ class SharedODLAPI(BaseCirculationAPI):
             if response.status_code == 404:
                 # 404 is returned when the loan has been deleted. Leave this loan out of the result.
                 continue
-            feed = feedparser.parse(str(response.content))
+            feed = self._parse_feed_from_response(response)
             entries = feed.get("entries")
             if len(entries) < 1:
                 raise CirculationException()
@@ -1354,7 +1444,7 @@ class SharedODLAPI(BaseCirculationAPI):
             if response.status_code == 404:
                 # 404 is returned when the hold has been deleted. Leave this hold out of the result.
                 continue
-            feed = feedparser.parse(str(response.content))
+            feed = self._parse_feed_from_response(response)
             entries = feed.get("entries")
             if len(entries) < 1:
                 raise CirculationException()
@@ -1518,3 +1608,41 @@ class MockSharedODLAPI(SharedODLAPI):
         self.request_args.append((patron, headers, allowed_response_codes))
         response = self.responses.pop()
         return HTTP._process_response(url, response, allowed_response_codes=allowed_response_codes)
+
+
+class ODLExpiredItemsReaper(IdentifierSweepMonitor):
+    """Responsible for removing expired ODL licenses."""
+
+    SERVICE_NAME = "ODL Expired Items Reaper"
+    PROTOCOL = ODLAPI.NAME
+
+    def __init__(self, _db, collection):
+        super(ODLExpiredItemsReaper, self).__init__(_db, collection)
+
+    def process_item(self, identifier):
+        for licensepool in identifier.licensed_through:
+            remaining_checkouts = 0   # total number of checkouts across all the licenses in the pool
+            concurrent_checkouts = 0  # number of concurrent checkouts allowed across all the licenses in the pool
+
+            # 0 is a starting point,
+            # we're going through all the valid licenses in the pool and count up available checkouts.
+            for license_pool_license in licensepool.licenses:
+                if not license_pool_license.is_expired:
+                    remaining_checkouts += license_pool_license.remaining_checkouts
+                    concurrent_checkouts += license_pool_license.concurrent_checkouts
+
+            if remaining_checkouts != licensepool.licenses_owned or \
+                    concurrent_checkouts != licensepool.licenses_available:
+                licenses_owned = max(remaining_checkouts, 0)
+                licenses_available = max(concurrent_checkouts, 0)
+
+                circulation_data = CirculationData(
+                    data_source=licensepool.data_source,
+                    primary_identifier=identifier,
+                    licenses_owned=licenses_owned,
+                    licenses_available=licenses_available,
+                    licenses_reserved=licensepool.licenses_reserved,
+                    patrons_in_hold_queue=licensepool.patrons_in_hold_queue,
+                )
+
+                circulation_data.apply(self._db, self.collection)